add results of first sys test
These first run results shows that the system is mostly working as
desired. However there are quite large errors in the timings, ie when
trying to setup for 72MHz the clock appears to run about 2MHz over and
the system clock appears jittery. This may be a hardware issue owing to
noise on the clock lines since they appear to run a long way on the pcb.

This will need further investigation.