add gpio driver to hal
A simple driver to read and write to gpio. This is the first driver
added to hal so a lot of the framework for drivers is going in to like
hw.h/hw.c etc.

Also a simple unit test to check the pins on the micro can be
set/cleared and read successfully, and finally a speed test of the GPIO
(about 580KHz atm) is done.