add slave spi module to hal layer
Added the spis (spi slave) module and unit tests. Currently the spi
reads work well, however the writes are hard to get consistent owing to
the double buffering. If testing with something like the nano board with
a MOSI phase, followed by a MISO phase you will see the MOSI data works
fine, however the MISO data can be shifted and get out of sync quite
easily for some reason. This _NEEDS_ more work.

It seems that certain changes (uart enable and sleep) are required to
sys_init() for the spi to function properly, I will look into why this
is in more detail later when I have time (I suspect the clock on the
blade DUT is jittery)