working sbc from 14/02/2013
this is the work we did on Wednesday night to get the b2s working
between the blade and the sbc. I am just making this branch as a lasting
stash of this working state, I will merge the good stuff across into the
working branches and hopefully only leave behind the debugging stuff.