add initial error handling to spis module
this just runs callbacks when an error is detected to let the caller
know to expect problems (ie check buffer more riggoursly or cancel the
io and start again)

the reads and writes keep running, and assuming enough clocks are
present will complete, it is just they may complete out of order.

there is a comment in spis about maybe changing the behaviour to
complete io ops in error, but atm I will leave this up to the
application as it allows more flexibility.