fix masive uninit var in spis bug
Initalise spis nvic irq to enabled. This was originally left as
uninitalised and hence the spis would sometimes work or not depending on
the values left in the memory by previous operations. This would also
explain all the optmisation issues (reordering of code would leave the
memory in different states 0 or not, hence enabling/disabling spis nvic
seemingly randomly)