dont check SPI_I2S_FLAG_TXE before the spis write
If a word from the last spis write is still stuck in the tx buffer then
just overwrite it and get on with this write. Previously we check the tx
empty flag and waited till it was clear before using the isr to start
sending this writes words, now we just over write the tx buffer
otherwise we will send 2 bytes (word) in front of this transfer.

NB: This is the cause of the b2s 2 byte shift issue. Somehow the
b2s_basic (etc) drops the select line in the middle of a transaction
and the b2s resets its state back to the read header state and starts a
new write, but because the tx buffer still has old data in it we will
forever be shifted by 2 bytes! The above fixes this by discarding the
stale word in the tx buffer.