initial crc stuff
add libs for dealing with crc's, a hal interface for both hardware and
software crc's, and a crc unit test.

@todo verify the software version of the crc matches the hardware
version and pc version.

@todo impliment crc8 and crc16

@todo use table method for software crc and drop crcmodel dependance
inside the hal crc.c. You will still need crcmodel to make the table,
but we can do the table calcs ourself since they dont exist in
crcmodel.c. Also we wil still need crctable.h as this is the table to
work with.