implement crc8 and crc16 (maybe ok?)
implement crc8 and crc16 as rounded most significant words of the crc32
algorithm found in the hardware.

I am not sure if this is a valid way to impliment crc8/crc16 but it will
be fast on the micro and I think it will provide reasonable error
detection.
see http://stackoverflow.com/questions/15199232/casting-crc32-to-uint8-t