makes the 8 bit tbl crc work for any cfg
This is really messy but it makes the 8 table bit crc work for any cm
config I try (ie when using table method compared to soft model).

Working is good, tidy is better so I need to fix this up (ideally wo
moding the crcmodel.c as this is an external file form the net).