make soft and table methods consistant
This is still not perfect, what I do is trick the lib into making the
table with refin as false, then we reflect is needed. Ideally the refin
setting can be built into the table, but it I having issues. This now
seems to be working, but we pay the price is it is slower.

Also I gave up on the refin setting used the flip the words, this should
always be true for ST lib which is just the default 8 bit access mode,
so this just comes down to endianness of the machine ?