first pass at getting a bootstrap up
this adds a hal interface to a bootstrap and also a mos bootstrap
program to check what programs are available and boot the best one.

the best programs are booted by the bootstrap in this order:
 1) any program requested by another program
 2) the primary program (if valid)
 3) the bootloader (if valid)
 4) give up and spin there looking sad for ourself :~( [I have no program]

so far the blade build has moved to accomodate the bootstrap, and a
bootstrap image is built as part of the makefile; but now we need to
merge these together [later w a bootloader prog] to make a final image,
and the test this.