hack to make spis_init work wo a prior gpio_init
for some reason the spis fails to work (irqs are dead at least and
possibly more) if gpio init is not called before doing the clk init for
the spi (spi3 at least).

this seems odd as gpio is on apb2 and spi3 is on apb1 so I wouldnt think
they should step on each others toes, but something is wrong.

as I dont have time to look into this more I will simply work around it
by calling the gpio_init for the spis pins (nss, miso, mosi, clk) before
the call to init the spis clk

I will come back to look at this when more time is available and make a
proper fix so it is not sensative to the order in which gpio_init and
spi_clk_init are called.

TLSX-33