add key and program max_len to program header
the key is needed to unlock this section so so the bootloader can over
write it

the max_len is the actual size allocated for the program (in bytes) so
the bootloader when unlocked will allow the caller to erase and fill
from the unlocked header to + max_len bytes. this is used to ensure the
bootloader only upgrades applications it is supposed to.