implement new packet protocol to master
the new protocol changes the size of some elements in the packet and
adds a pause followed by a ack/nak to allow the host and slave to let
each other know if the crc passed on the packets they send:
  <payload[62 bytes]><status [2 bytes]><crc [4 bytes]> ...
  {master dead time} ... <ack[2 bytes]>

the master transfers 62 payload bytes followed by 2 status bytes,
followed by 4 crc bytes. then the master inserts a pause to allow both
sides to calculate the crc on the packets they received, then the rx
packets of each side are either ACK'ed or NAK'ed to allow each side to
know if they need to resend the tx packet.

although these sizes seem to be overkill, it seems to be needed to align
the crc to 4 bytes (otherwise the last bytes in the crc will be randomly
different on mater and slave). also the spi must transfer 2 bytes each
time so there must be multiples of 2 at the pause, and after, or the
bytes will be out of sync.

a simple state machine was added to manage the process of transmitting
and receiving the different parts of the packets. (@todo document the
state machine)