add basic support for new STM32F373 chip
this adds the new stm32f373 chip into the HAL (our second chip, yay) but
it is currently missing spis and bootstrap modules as these are not
tested yet.

also this patch updates the unit tests for the modules that are done so
they work between all the chips (some where changed to use the
development board a bit more too, but should still work on the
STM32f107x)

@todo - add and test bootstrap module and spis module (then this will be
functionally equivilant to the stm32f107x and then we add more driver we
want).