add spis support for stm32f373
This adds untested support for the slave spi on the stm32f373 chip. This
is likely very buggy at this stage and will need work. It is based on
the spis from the stm32f107x (which is not great either), with some
changes to support the new API and some improvements to flush the
buffers (now they are bigger it is much easier to get out of sync!)