clean up on spi tx flush reseting
Since st did not provide a way to flush the spi tx buffer, we must reset
the peripheral (we do this via SPI_I2S_DeInit now). But this leaves the
spi registers in their reset state. Previously we re ran the whole
spis_init again, but this was slow and since we might do this often (ie
on select/deselect) I call only what is needed now in spis_init_regs()
which is called from both spis_init and the spi_flush_tx_fifo() routines
so we don't duplicate code.