fix build (bootstrap now fits)
Allow some hal modules to not be linked into the final hal lib. A config
file is used to control which optional modules are included in the final
library, ie to include GPIO put a line in the config file like:
CONFIG_GPIO = y
This was done to save space in the final binary since the optional
modules define ISR's and gcc will assume they might run, hence the only
way to save the space if these modules are not used is to not link them.

If no config file is provided then just the bare minimum files (the non
optional ones) are included as though all module configs were set = n

We also make the name of libhal a variable so it can be changed for
example if you have an application that needs a complete libhal and a
application that only needs the bare minimum, then the linker could make
a libhal_full.o and a libhal_min.o. Also this variable is added to the
.PHONY list in the make file so the linking will always happen since you
might be using a new config file

for the bootstrap we do what is mentioned above, we build a libhal_bs.o
that has just enough for the bootstrap and the bootstrap links against
this, but the unit tests link against libhal.o which they rebuild as
they go for their needs.