add support for spi master
atm this is a big mess and I need to split the master and slave stuff
etc, but it works and the master and slave can talk to each other

I could not get the hardware NSS support on the stm32 working so I have
left it as a normal gpio pin. As a consquence of that I need to enable
the spi before writing into the fifo so, then set the NSS line low so
the clock line is set hi or lo (as configured in hw.c) when before the
NSS line goes low; this is a pain since I cannot stuff the Tx fifo with
bytes before enabling the spi which would be nice.

Otherwise this seems to work OK